pyproject.toml
[project]
name = "datenna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.9.0.post1",
    "ipykernel>=6.29.5",
    "langdetect>=1.0.9",
    "matplotlib>=3.9.3",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "sentence-transformers>=3.3.1",
]
